requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
cssselect2==0.7.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
langdetect==1.0.9
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
openai==1.9.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
validators==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
